added test for get_reader and moved it to file_handling module since it works with files
